requirements.txt
Adafruit-Blinka==8.31.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-requests==2.0.4
adafruit-circuitpython-rgb-display==3.12.3
adafruit-circuitpython-typing==1.9.6
Adafruit-GPIO==1.0.3
adafruit-io==2.7.1
Adafruit-PlatformDetect==3.59.0
Adafruit-PureIO==1.1.11
board==1.0
certifi==2023.11.17
charset-normalizer==3.3.2
get-mac==0.9.2
idna==3.6
modbus-tk==1.1.3
paho-mqtt==1.6.1
pillow==10.2.0
psutil==5.9.8
pyftdi==0.55.0
pypng==0.20220715.0
pyserial==3.5
pyusb==1.2.1
qrcode==7.4.2
requests==2.31.0
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
spidev==3.6
sysv_ipc==1.1.0
typing_extensions==4.9.0
urllib3==2.1.0
